refactor: split navigation service and UI
