+++ src/helpers/navigation/navData.js
@@ -1,22 +1,6 @@
-import { debugLog } from "../debug.js";
 import { loadSettings } from "../settingsStorage.js";
 import { loadNavigationItems } from "../gameModeUtils.js";
-
-/**
- * Validates game modes data to ensure all required properties are present.
- *
- * @pseudocode
- * 1. Filter out items missing `name` or `url` properties.
- * 2. Return the filtered array.
- *
- * @param {import("../types.js").GameMode[]} gameModes - The list of game modes to validate.
- * @returns {import("../types.js").GameMode[]} Validated game modes.
- */
-export function validateGameModes(gameModes) {
-  const validatedModes = gameModes.filter((mode) => mode.name && mode.url);
-  debugLog("Validated modes:", validatedModes); // Debug validated modes
-  return validatedModes;
-}
+import { validateGameModes } from "./navigationService.js";
 
 /**
  * Load and validate main menu navigation items, marking hidden modes based on
