+++ src/helpers/navigation/navigationUI.js
@@ -1,53 +1,4 @@
-import { validateGameModes } from "../navigation/navData.js";
-
-/**
- * Convert a game mode name to a camelCase key for tooltip ids.
- *
- * @pseudocode
- * 1. Replace non-alphanumeric characters with spaces.
- * 2. Split into words.
- * 3. Lowercase the first word and capitalize the rest.
- * 4. Join the words without spaces.
- *
- * @param {string} name - Game mode name.
- * @returns {string} camelCase key.
- */
-export function navTooltipKey(name) {
-  return String(name)
-    .replace(/[^a-zA-Z0-9]+/g, " ")
-    .trim()
-    .split(/\s+/)
-    .map((w, i) => (i === 0 ? w.toLowerCase() : w[0].toUpperCase() + w.slice(1)))
-    .join("");
-}
-
-/**
- * Escape special HTML characters to prevent injection.
- *
- * @pseudocode
- * 1. Convert the input to a string.
- * 2. Replace `&`, `<`, `>`, `"`, and `'` with their HTML entity equivalents.
- *
- * @param {unknown} value - The value to escape.
- * @returns {string} Escaped string safe for HTML insertion.
- */
-export function escapeHtml(value) {
-  return String(value)
-    .replace(/&/g, "&amp;")
-    .replace(/</g, "&lt;")
-    .replace(/>/g, "&gt;")
-    .replace(/"/g, "&quot;")
-    .replace(/'/g, "&#39;");
-}
-
-/**
- * Base path for navigation links derived from the current module location.
- */
-export const BASE_PATH = (() => {
-  const url = new URL(import.meta.url);
-  url.pathname = url.pathname.replace(/helpers\/api\/navigation.js$/, "pages/");
-  return url;
-})();
+import { validateGameModes, navTooltipKey, escapeHtml, BASE_PATH } from "./navigationService.js";
 
 /**
  * Clears existing content in the bottom navigation bar.
