+++ src/helpers/setupBottomNavbar.js
@@ -3,7 +3,7 @@
  *
  * @pseudocode
  * 1. Import `loadMenuModes` from `navigation/navData.js`.
- * 2. Import `buildMenu` and `setupHamburger` from `api/navigation.js`.
+ * 2. Import `buildMenu` and `setupHamburger` from `navigation/navigationUI.js`.
  * 3. Import `populateNavbar` from `navigationBar.js`.
  * 4. Import `setupButtonEffects` from `buttonEffects.js`.
  * 5. Import `onDomReady`.
@@ -12,7 +12,7 @@
  * 8. Use `onDomReady` to invoke `init` with guarded error handling.
  */
 import { loadMenuModes } from "./navigation/navData.js";
-import { buildMenu, setupHamburger } from "./api/navigation.js";
+import { buildMenu, setupHamburger } from "./navigation/navigationUI.js";
 import { populateNavbar } from "./navigationBar.js";
 import { setupButtonEffects } from "./buttonEffects.js";
 import { onDomReady } from "./domReady.js";
